metadata: add support for publisher from metadata

+++ upnpsoap.h
@@ -40,7 +40,7 @@ struct Response
 	int returned;
 	int requested;
 	int iface;
-	uint32_t filter;
+	uint64_t filter;
 	uint32_t flags;
 	enum client_types client;
 };
